refactor(useform): replace old implementation
