+++ example/src/Controlled/index.tsx
@@ -1,6 +1,6 @@
 import { TextField } from '@material-ui/core'
 import * as React from 'react'
-import { useForm } from '../../../src/hooks/hook'
+import { useForm } from '../../../src/hooks/useForm'
 import Select from 'react-select'
 import * as yup from 'yup'
